refactor to extract token modal logic to a single file
